solicitud-service/requirements.txt
flask
pyjwt
requests
zeep
spyne
lxml
